+++ frontend/src/ApplicationLayout/Navigation/SyllabusNav.tsx
