+++ frontend/src/screens/CoursePage/CourseButtonBar.tsx
