+++ frontend/src/screens/CoursePage/CourseCard.css
