+++ frontend/src/screens/CoursePage/CourseCard.tsx
