+++ frontend/src/screens/CoursePage/CoursePage.css
