+++ frontend/src/screens/CoursePage/CoursePage.tsx
