+++ frontend/src/screens/SyllabusView/BasicInformation/BasicInfo.css
@@ -1,4 +1,4 @@
-.basic-info-container {
+.course-info-container {
     max-width: 1200px;
     margin: 2rem auto;
     padding: 2rem;
@@ -20,20 +20,3 @@
     font-size: 1.2rem;
 }
 
-
-.save-button {
-    display: inline-block;
-    margin-top: 1rem;
-    padding: 0.75rem 1.5rem;
-    background-color: #002e5d;
-    color: #fff;
-    border: none;
-    border-radius: 4px;
-    font-weight: bold;
-    cursor: pointer;
-    transition: background-color 0.2s ease;
-}
-
-.save-button:hover {
-    background-color: #004080;
-}
